Undo placing cached jar with Paperclip
This is no longer needed, as both Java 8 and Java 9 implementations
handle adding classpath entries properly, without requiring on fragile
jar manifest entries. While it is not needed anymore, keeping the cache
directory with the Paperclip jar could cause several issues, especially:

 * In cases where the user does not have write access to the Paperclip
   jar itself, such as a hosted server.

 * In cases where a user is starting up multiple Paperclip servers on
   the same machine using the same Paperclip jar at the same time.

I don't imagine the second point to actually be a common problem, but it
is technically an issue.